[#121200693]: Add new test file , make the u values as explicit inteers and add missing modulo

+++ anoncreds/protocol/prover.py
@@ -253,7 +253,7 @@ class Prover:
 
 
 def findLargestSquareLessThan(x):
-    sqrtx = floor(sqrt(x))
+    sqrtx = int(floor(sqrt(x)))
     return sqrtx
